Tools: allow IAT verification with public key
Due to an issue with pycose library, signature verification needed to be
done using the signing (private) key. The fix for this issue has been
recently merged:

        https://github.com/TimothyClaeys/COSE-PYTHON/commit/64cea44a4a81dd8b02a86eae8798962d4afbb148

This allows the use of the verifying (public) key to verify signatures.

This commit updates iat-verifier to enable it to use either key for
signature verification.

Change-Id: Ia4b228adec8b2b56935af1737f6f1bd99302959b
Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>